backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
python-dotenv==1.0.0
pydantic==2.5.0
slowapi==0.1.9
redis==5.0.1
squareup==39.1.0.20241218
boto3==1.34.0

# Authentication dependencies
alembic==1.13.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.1.2
email-validator==2.1.0
pyjwt==2.8.0

# Email service (optional - choose one)
# sendgrid==6.11.0  # Alternative to AWS SES
# Note: boto3 already included above for AWS SES

# ML/AI dependencies for anomaly detection and forecasting
scikit-learn>=1.4.0
numpy>=1.26.0
pandas>=2.2.0
statsmodels>=0.14.0
requests>=2.31.0
